Fixes for server
Fix cursor position on FF, where scenario container was slightly further downpage.  Now the toolboxes are floating so that the map tiles are always in the same spot.

Fix cursor event buffering so that it reschedules the next buffer even if the WS wasn't ready this time.  Fixes the issue where cursor wouldn't work if you moved the mouse before WS had fully established.

Fix incremental state tracking so that the correct "last event" UUID is given whenever the full state is fetched.  Previously, between the collapsing of redundant events, and reordering on load (now removed), the last event according to the server would sometimes not be the last one processed on the client.